refactor: move snap placement to schematics/snap.py
Snap geometry functions (snap_two_pin_parts, stagger, pre-shift) are
backend-agnostic — they operate on Point/Tx/orientation strings, not
KiCad S-expressions. Moved to schematics/ per maintainer feedback.

KiCad-specific sexp output (wireable nets, power bus wires, no-connect
flags) stays in tools/kicad9/sexp_schematic.py.
