+++ src/skidl/tools/kicad9/gen_schematic.py
@@ -13,12 +13,14 @@ import os
 import re
 import shutil
 import subprocess
-from collections import Counter, defaultdict
+from collections import Counter
 
 from skidl.geometry import BBox, Point, Tx, Vector
 from skidl.schematics.net_terminal import NetTerminal
 from skidl.scriptinfo import get_script_name
 from skidl.utilities import export_to_all, rmv_attr
+
+from skidl.schematics.snap import snap_two_pin_parts
 
 from .sexp_schematic import write_top_schematic
 from .bboxes import calc_hier_label_bbox, calc_symbol_bbox
@@ -348,450 +350,6 @@ def _handle_fallback(circuit, tool_module, filepath, top_name, title, flatness,
         warnings.warn(msg, LabelsOnlyWarning, stacklevel=4)
 
 
-# ---------------------------------------------------------------------------
-# Snap placement — move 2-pin parts onto IC pins for cleaner schematics
-# ---------------------------------------------------------------------------
-
-
-def _is_two_pin_part(part):
-    """Return True if part is a simple 2-pin component (LED, R, C, etc.)."""
-    return not isinstance(part, NetTerminal) and len(part.pins) == 2
-
-
-def _is_power_net(net):
-    """Return True if net is a power rail (GND, VCC, +3.3V, etc.)."""
-    name = getattr(net, 'name', '')
-    return name.startswith("+") or bool(_POWER_NET_RE.match(name))
-
-
-def _pin_world_orient(pin, part):
-    """Get the world-space outward direction from a pin after part rotation."""
-    orient_to_vec = {"R": (1, 0), "L": (-1, 0), "U": (0, -1), "D": (0, 1)}
-    outward = {"L": "R", "R": "L", "U": "D", "D": "U"}
-
-    raw_orient = getattr(pin, "orientation", "R")
-    vx, vy = orient_to_vec.get(raw_orient, (1, 0))
-    tx = part.tx
-    wx = tx.a * vx + tx.b * vy
-    wy = tx.c * vx + tx.d * vy
-    if abs(wx) >= abs(wy):
-        world_orient = "R" if wx > 0 else "L"
-    else:
-        world_orient = "D" if wy > 0 else "U"
-    return outward.get(world_orient, "R")
-
-
-def _compute_snap_tx(my_pin, other_pin, target_world, extend_dir):
-    """Compute the transform to snap a 2-pin part onto a target pin position."""
-    dx_local = other_pin.pt.x - my_pin.pt.x
-    dy_local = other_pin.pt.y - my_pin.pt.y
-
-    if extend_dir == "R":
-        if abs(dx_local) >= abs(dy_local):
-            symtx = "" if dx_local > 0 else "H"
-        else:
-            symtx = "R" if dy_local > 0 else "L"
-    elif extend_dir == "L":
-        if abs(dx_local) >= abs(dy_local):
-            symtx = "" if dx_local < 0 else "H"
-        else:
-            symtx = "L" if dy_local > 0 else "R"
-    elif extend_dir == "U":
-        if abs(dy_local) >= abs(dx_local):
-            symtx = "" if dy_local > 0 else "V"
-        else:
-            symtx = "L" if dx_local > 0 else "R"
-    elif extend_dir == "D":
-        if abs(dy_local) >= abs(dx_local):
-            symtx = "" if dy_local < 0 else "V"
-        else:
-            symtx = "R" if dx_local > 0 else "L"
-    else:
-        symtx = ""
-
-    new_tx = Tx.from_symtx(symtx)
-    my_pin_placed = my_pin.pt * new_tx
-    offset = Point(
-        target_world.x - my_pin_placed.x,
-        target_world.y - my_pin_placed.y,
-    )
-    return new_tx.move(offset)
-
-
-def _snap_two_pin_parts(node):
-    """Snap 2-pin parts onto their connected IC pins post-route.
-
-    Pass 1: Snap onto IC pins (parts with >2 pins).
-    Pass 2+: Chain onto already-snapped 2-pin parts.
-    Pass 3: Stack remaining onto occupied IC pins perpendicular.
-    """
-    for child in node.children.values():
-        _snap_two_pin_parts(child)
-
-    node_part_ids = {id(p) for p in node.parts}
-    snapped = set()
-    occupied_pins = set()
-
-    for part in list(node.parts):
-        if not _is_two_pin_part(part):
-            continue
-
-        p1, p2 = part.pins[0], part.pins[1]
-        net1 = getattr(p1, "net", None)
-        net2 = getattr(p2, "net", None)
-        if not net1 or not net2:
-            continue
-
-        target_pin = None
-        target_part = None
-        my_pin = None
-
-        both_power = _is_power_net(net1) and _is_power_net(net2)
-        min_target_pins = 8 if both_power else 2
-
-        for my_p, other_net in [(p1, net1), (p2, net2)]:
-            if _is_power_net(other_net) and not both_power:
-                continue
-            for net_pin in other_net.pins:
-                other_part = net_pin.part
-                if (
-                    other_part is not part
-                    and id(other_part) in node_part_ids
-                    and not isinstance(other_part, NetTerminal)
-                    and len(other_part.pins) > min_target_pins
-                    and id(net_pin) not in occupied_pins
-                ):
-                    target_pin = net_pin
-                    target_part = other_part
-                    my_pin = my_p
-                    break
-            if target_pin:
-                break
-
-        if not target_pin:
-            continue
-
-        target_world = target_pin.pt * target_part.tx
-        extend_dir = _pin_world_orient(target_pin, target_part)
-        other_pin = p2 if my_pin is p1 else p1
-
-        part.tx = _compute_snap_tx(my_pin, other_pin, target_world, extend_dir)
-        if both_power:
-            _offset_dir = {"R": (200, 0), "L": (-200, 0), "U": (0, 200), "D": (0, -200)}
-            dx, dy = _offset_dir.get(extend_dir, (200, 0))
-            part.tx = part.tx.move(Point(dx, dy))
-            cap_pin_world = my_pin.pt * part.tx
-            power_cap_wires = getattr(node, "_power_cap_wires", [])
-            power_cap_wires.append(
-                (target_world.x, target_world.y, cap_pin_world.x, cap_pin_world.y)
-            )
-            node._power_cap_wires = power_cap_wires
-            power_cap_suppressed = getattr(node, "_power_cap_suppressed_pins", set())
-            power_cap_suppressed.add(id(my_pin))
-            node._power_cap_suppressed_pins = power_cap_suppressed
-        snapped.add(id(part))
-        occupied_pins.add(id(target_pin))
-
-    for _iteration in range(5):
-        newly_snapped = set()
-
-        for part in list(node.parts):
-            if id(part) in snapped or not _is_two_pin_part(part):
-                continue
-
-            p1, p2 = part.pins[0], part.pins[1]
-            net1 = getattr(p1, "net", None)
-            net2 = getattr(p2, "net", None)
-            if not net1 or not net2:
-                continue
-
-            target_pin = None
-            target_part = None
-            my_pin = None
-
-            both_power = _is_power_net(net1) and _is_power_net(net2)
-
-            for my_p, other_net in [(p1, net1), (p2, net2)]:
-                if _is_power_net(other_net) and not both_power:
-                    continue
-                for net_pin in other_net.pins:
-                    other_part = net_pin.part
-                    if (
-                        other_part is not part
-                        and id(other_part) in snapped
-                        and id(net_pin) not in occupied_pins
-                    ):
-                        target_pin = net_pin
-                        target_part = other_part
-                        my_pin = my_p
-                        break
-                if target_pin:
-                    break
-
-            if not target_pin:
-                continue
-
-            target_world = target_pin.pt * target_part.tx
-            extend_dir = _pin_world_orient(target_pin, target_part)
-            other_pin = p2 if my_pin is p1 else p1
-
-            part.tx = _compute_snap_tx(my_pin, other_pin, target_world, extend_dir)
-            newly_snapped.add(id(part))
-            occupied_pins.add(id(target_pin))
-
-        if not newly_snapped:
-            break
-        snapped |= newly_snapped
-
-    perp_map = {"R": "D", "L": "U", "U": "R", "D": "L"}
-    for part in list(node.parts):
-        if id(part) in snapped or not _is_two_pin_part(part):
-            continue
-
-        p1, p2 = part.pins[0], part.pins[1]
-        net1 = getattr(p1, "net", None)
-        net2 = getattr(p2, "net", None)
-        if not net1 or not net2:
-            continue
-
-        target_pin = None
-        target_part = None
-        my_pin = None
-
-        for my_p, other_net in [(p1, net1), (p2, net2)]:
-            for net_pin in other_net.pins:
-                other_part = net_pin.part
-                if (
-                    other_part is not part
-                    and id(other_part) in node_part_ids
-                    and not isinstance(other_part, NetTerminal)
-                    and len(other_part.pins) > 2
-                    and id(net_pin) in occupied_pins
-                ):
-                    target_pin = net_pin
-                    target_part = other_part
-                    my_pin = my_p
-                    break
-            if target_pin:
-                break
-
-        if not target_pin:
-            continue
-
-        target_world = target_pin.pt * target_part.tx
-        ic_dir = _pin_world_orient(target_pin, target_part)
-        extend_dir = perp_map.get(ic_dir, ic_dir)
-        other_pin = p2 if my_pin is p1 else p1
-
-        part.tx = _compute_snap_tx(my_pin, other_pin, target_world, extend_dir)
-        snapped.add(id(part))
-
-    _stagger_tjunctions(node, node_part_ids, snapped, occupied_pins)
-
-
-def _stagger_tjunctions(node, node_part_ids, snapped, occupied_pins, min_group=2):
-    """Detect repeating T-junction patterns and stagger parts outward from IC."""
-    perp_map = {"R": "D", "L": "U", "U": "R", "D": "L"}
-    anti_perp = {"U": "D", "D": "U", "L": "R", "R": "L"}
-    _dir_vec = {"R": (1, 0), "L": (-1, 0), "U": (0, -1), "D": (0, 1)}
-
-    ic_pin_to_parts = defaultdict(list)
-
-    for part in node.parts:
-        if not _is_two_pin_part(part):
-            continue
-
-        p1, p2 = part.pins[0], part.pins[1]
-        net1 = getattr(p1, "net", None)
-        net2 = getattr(p2, "net", None)
-        if not net1 or not net2:
-            continue
-
-        for my_p, other_net in [(p1, net1), (p2, net2)]:
-            if _is_power_net(other_net):
-                continue
-            for net_pin in other_net.pins:
-                ic = net_pin.part
-                if (
-                    ic is not part
-                    and id(ic) in node_part_ids
-                    and not isinstance(ic, NetTerminal)
-                    and len(ic.pins) > 2
-                    and id(net_pin) in occupied_pins
-                ):
-                    other_pin = p2 if my_p is p1 else p1
-                    ic_pin_to_parts[id(net_pin)].append(
-                        (part, my_p, other_pin, net_pin, ic)
-                    )
-                    break
-            else:
-                continue
-            break
-
-    ic_groups = defaultdict(list)
-    for ic_pin_id, parts_list in ic_pin_to_parts.items():
-        if not parts_list:
-            continue
-        ic = parts_list[0][4]
-        ic_groups[id(ic)].append((parts_list[0][3], parts_list))
-
-    MM_TO_MILS = 1 / 0.0254
-    stagger_plans = []
-
-    for ic_id, pin_entries in ic_groups.items():
-        fanout_counts = [len(pl) for _, pl in pin_entries]
-        dominant = max(set(fanout_counts), key=fanout_counts.count)
-        if dominant < 2:
-            continue
-        matching = [(ip, pl) for ip, pl in pin_entries if len(pl) == dominant]
-
-        if len(matching) < min_group:
-            continue
-
-        ic_part = matching[0][1][0][4]
-        ic_dir = _pin_world_orient(matching[0][0], ic_part)
-        step_dx, step_dy = _dir_vec.get(ic_dir, (1, 0))
-
-        max_span = 0
-        for _, parts_list_scan in matching:
-            for (scan_part, _, _, _, _) in parts_list_scan:
-                pts = [getattr(p, "pt", Point(p.x * MM_TO_MILS, p.y * MM_TO_MILS)) for p in scan_part.pins]
-                if pts:
-                    span = max(
-                        max(p.x for p in pts) - min(p.x for p in pts),
-                        max(p.y for p in pts) - min(p.y for p in pts),
-                    )
-                    max_span = max(max_span, span)
-        step_size = max(100, int(max_span) + 50)
-
-        n_pins = len(matching)
-        stagger_extent = step_size * n_pins + max_span
-
-        stagger_plans.append({
-            "ic_part": ic_part,
-            "matching": matching,
-            "ic_dir": ic_dir,
-            "step_dx": step_dx,
-            "step_dy": step_dy,
-            "step_size": step_size,
-            "stagger_extent": stagger_extent,
-            "dominant": dominant,
-        })
-
-    if len(stagger_plans) > 1:
-        _pre_shift_ics(stagger_plans, node, snapped)
-
-    junction_wires = getattr(node, "_tjunction_wires", [])
-    suppressed_pins = set()
-
-    for plan in stagger_plans:
-        ic_part = plan["ic_part"]
-        matching = plan["matching"]
-        ic_dir = plan["ic_dir"]
-        step_dx = plan["step_dx"]
-        step_dy = plan["step_dy"]
-        step_size = plan["step_size"]
-        perp_dir = perp_map.get(ic_dir, ic_dir)
-
-        def _pin_sort_key(entry, _ic_part=ic_part, _ic_dir=ic_dir):
-            ic_pin = entry[0]
-            w = ic_pin.pt * _ic_part.tx
-            if _ic_dir in ("L", "R"):
-                return w.y
-            return w.x
-
-        matching.sort(key=_pin_sort_key)
-
-        parts_per_pin = plan["dominant"]
-        anti = anti_perp.get(perp_dir, perp_dir)
-        extend_dirs = [perp_dir, anti] if parts_per_pin >= 2 else [perp_dir]
-
-        for pin_idx, (ic_pin, parts_list) in enumerate(matching):
-            ic_pin_world = ic_pin.pt * ic_part.tx
-
-            parts_list.sort(key=lambda t: getattr(t[0], "ref", ""))
-
-            offset_n = pin_idx + 1
-            ox = ic_pin_world.x + step_dx * step_size * offset_n
-            oy = ic_pin_world.y + step_dy * step_size * offset_n
-            junction_pt = Point(ox, oy)
-
-            for part_idx, (part, my_pin, other_pin, _, _) in enumerate(parts_list):
-                ext_dir = extend_dirs[part_idx % len(extend_dirs)]
-                part.tx = _compute_snap_tx(
-                    my_pin, other_pin, junction_pt, ext_dir
-                )
-                snapped.add(id(part))
-                suppressed_pins.add(id(my_pin))
-            junction_wires.append(
-                (ic_pin_world.x, ic_pin_world.y, ox, oy)
-            )
-
-    node._tjunction_wires = junction_wires
-    node._tjunction_suppressed_pins = suppressed_pins
-
-
-def _pre_shift_ics(plans, node, snapped):
-    """Shift ICs vertically so stagger fans won't overlap."""
-    for plan in plans:
-        ic = plan["ic_part"]
-        ic_deps = set()
-        ic_id = id(ic)
-
-        for part in node.parts:
-            if id(part) == ic_id or id(part) not in snapped:
-                continue
-            if not _is_two_pin_part(part):
-                continue
-            for pin in part.pins:
-                net = getattr(pin, "net", None)
-                if not net:
-                    continue
-                for net_pin in net.pins:
-                    if net_pin.part is ic:
-                        ic_deps.add(id(part))
-                        break
-                if id(part) in ic_deps:
-                    break
-
-        plan["_deps"] = [p for p in node.parts if id(p) in ic_deps]
-
-    def _ic_bbox(plan):
-        ic = plan["ic_part"]
-        all_parts = [ic] + plan["_deps"]
-        min_y = float("inf")
-        max_y = float("-inf")
-        for part in all_parts:
-            for pin in part.pins:
-                w = pin.pt * part.tx
-                min_y = min(min_y, w.y)
-                max_y = max(max_y, w.y)
-        return min_y, max_y
-
-    plans.sort(key=lambda p: _ic_bbox(p)[0])
-
-    margin = 200
-    prev_max_y = None
-
-    for plan in plans:
-        ic_min_y, ic_max_y = _ic_bbox(plan)
-        needed_height = plan["stagger_extent"]
-        group_max_y = max(ic_max_y, ic_min_y + needed_height)
-
-        if prev_max_y is not None and ic_min_y < prev_max_y + margin:
-            shift = (prev_max_y + margin) - ic_min_y
-            vec = Point(0, shift)
-            shifted = set()
-            for part in [plan["ic_part"]] + plan["_deps"]:
-                if id(part) not in shifted:
-                    part.tx = part.tx.move(vec)
-                    shifted.add(id(part))
-            ic_min_y += shift
-            group_max_y += shift
-
-        prev_max_y = group_max_y
-
-
 def _stub_all_non_explicit(circuit):
     """Stub all nets that weren't explicitly set by the user (labels-only fallback).
 
@@ -1058,7 +616,7 @@ def gen_schematic(
             continue
 
         if options.get("auto_stub", False):
-            _snap_two_pin_parts(node)
+            snap_two_pin_parts(node)
 
         output_file = write_top_schematic(
             circuit, node, filepath, top_name, title, version=20230409
@@ -1110,7 +668,7 @@ def gen_schematic(
                             _classify_and_stub_complex_nets(circuit, node, **options)
                         node.route(**options)
                         if options.get("auto_stub", False):
-                            _snap_two_pin_parts(node)
+                            snap_two_pin_parts(node)
                         output_file = write_top_schematic(
                             circuit, node, filepath, top_name, title, version=20230409
                         )
